Add files via upload
We implement the forward longitudinal vehicle model. The model accepts throttle inputs and steps through the longitudinal dynamic equations. Once implemented, you will be given a set of inputs that drives over a small road slope to test your model.

We set the longitudinal model inside a Python class. The vehicle begins to move with an initial velocity of 5 m/s and engine speed of 100 rad/s. All the relevant parameters are defined and a sampling time of 10ms is used for numerical integration. We implement the combined engine dynamic equations along with the force equations.